Variable selection method for GLMMs

A variable selection method for generalize linear mixed-effect models.

TODO: replace with the project's funding

Much of this package has been developed to support the needs of the following projects. We appreciate the support.

OUHSC CCAN Independent Evaluation of the State of Oklahoma Competitive Maternal, Infant, and Early Childhood Home Visiting (MIECHV) Project. HRSA/ACF D89MC23154. David Bard, PI, OUHSC; 2011-2015.

Independent Evaluation of the State of OK MIECHV Evidence Based Home Visitation Project, NIH-sponsored collaboration with OSDH. David Bard, PI, OUHSC; 2015-2017.

OSDH ParentPRO Pilot Evaluation, federally-sponsored collaboration with OSDH. David Bard, PI, OUHSC; 2015-2017.

Title IV-E Waiver Project, HRSA/MCHB-sponsored collaboration with OKDHS; David Bard, PI, OUHSC; 2014-2017.

Integrative Analysis of Longitudinal Studies of Aging (IALSA), sponsored by NIH 5P01AG043362. Scott Hofer, PI, University of Victoria; Will Beasley, PI of site-award, OUHSC; 2013-2018.

Oklahoma Shared Clinical and Translational Resources, sponsored by NIH NIGMS; U54 GM104938. Judith A. James, PI, OUHSC; 2013-2018.

Additional Institutional Support from OUHSC Dept of Pediatrics; 2013-2017.

Note

The release version is not yet available. It will be available through CRAN by running install.packages('RandomRPQL'). The most recent development version is available through GitHub by running remotes::install_github('OuhscBbmc/RandomRPQL') (make sure remotes is already installed). If you're having trouble with the package, please install the development version. If this doesn't solve your problem, please create a new issue, or email Will.

Examples

# \dontrun{
# Install/update RandomRPQL with the release version from CRAN.
# install.packages('RandomRPQL')

# Install/update RandomRPQL with the development version from GitHub
# install.packages("remotes") # Uncomment if `remotes` isn't installed already.
# remotes::install_github('OuhscBbmc/RandomRPQL')
# }